requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai>=1.52.0
PyPDF2==3.0.1
pdfplumber==0.11.0
pymupdf==1.24.5
python-docx==1.1.2
requests==2.32.3
pillow==10.3.0
anthropic>=0.34.0
PyJWT==2.8.0
gunicorn==22.0.0
